CleanVessel/components/VesselSidebar/VesselReportingList from old shared props
